Adding the actual script
This file scrapes the mobile Kijiji website indicated in the getWebsiteData function. It searches for and excludes given strings in the getListingsData function and then compares the listings to a dump file created by the program. New listings are displayed in the console along with their price, and all listings found for the page are stored in a text file and a dump file.